fix week2 code

+++ week2/code/basic_io2.py
@@ -2,6 +2,13 @@
 # FILE OUTPUT
 #############################
 # Save the elements of a list to a file
+import os
+import sys
+
+if not os.path.exists('../sandbox/test.txt'): ## make sure the folder exists
+    print("not found: ../sandbox/test.txt")
+    sys.exit(1)
+    
 list_to_save = range(100)
 
 f = open('../sandbox/testout.txt','w')
